feat: coutdown widget added

+++ lib/feature/auth/data/model/send_otp_response.dart
@@ -10,6 +10,6 @@ class SendOtpResponse {
   factory SendOtpResponse.fromJson(Map<String, dynamic> json) =>
       SendOtpResponse(
         otpToken: json["otpToken"],
-        cooldown: json["cooldown"],
+        cooldown: json["cooldown"].toDouble(),
       );
 }
